Time: 45 ms (20.88%), Space: 43.4 MB (36.63%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -22,7 +22,7 @@ class Solution {
         return res;
     }
     
-    private int eat(double speed,int[] piles)
+    private int eat(int speed,int[] piles)
     {
         
         int hrsNeeded = 0;
